Move commands and console variable into s_sound.c
